requirements.txt
# Enterprise RAG System - Python Dependencies
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
psutil==5.9.6
pyyaml==6.0.1

# Vector database
qdrant-client==1.7.0

# LLM and AI
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
bitsandbytes==0.41.3
peft==0.7.1

# Document processing
python-magic==0.4.27
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# Text processing
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 